requirements.txt
streamlit>=1.36
requests>=2.31
python-dotenv>=1.0.1
openai>=1.40.0
pydantic>=2.7
tqdm>=4.66
pypdf>=4.2.0
flask
python-dotenv
requests
